fixing

+++ alymova.ksenia/T1/regular.hpp
@@ -19,7 +19,7 @@ namespace alymova
   private:
     point_t pos_, top_, other_;
     double radius_big_, radius_small_, other_side_;
-    int sides_cnt_;
+    size_t sides_cnt_;
     rectangle_t frame_rect_;
   };
 }
